+++ src/main/java/br/com/bank/domain/domain/LimitTransaction.java
@@ -4,9 +4,11 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
 public class LimitTransaction {
-    private Integer limitPerTransaction;
+    private BigDecimal limitPerTransaction;
 }
